fix: abstracted the bids & asks calculation
